requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via aiohttp
av==15.1.0
    # via -r requirements.in
bencode2==0.3.26
    # via -r requirements.in
bitarray==3.7.1
    # via bitstring
bitstring==4.3.1
    # via -r requirements.in
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
click==8.2.1
    # via uvicorn
coverage[toml]==7.10.6
    # via pytest-cov
fastapi==0.116.1
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via respx
idna==3.10
    # via
    #   anyio
    #   httpx
    #   yarl
iniconfig==2.1.0
    # via pytest
libtorrent==2.0.11
    # via -r requirements.in
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
numpy==2.3.2
    # via -r requirements.in
packaging==25.0
    # via pytest
pillow==11.3.0
    # via -r requirements.in
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.7
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
pygments==2.19.2
    # via pytest
pysqlite3-binary==0.5.4
    # via -r requirements.in
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==1.1.0
    # via -r requirements.in
pytest-cov==6.2.1
    # via -r requirements.in
python-dotenv==1.1.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via uvicorn
respx==0.22.0
    # via -r requirements.in
six==1.17.0
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via -r requirements.in
starlette==0.47.3
    # via fastapi
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   bencode2
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
uvicorn[standard]==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
yarl==1.20.1
    # via aiohttp
